+++ src/TortoiseProc/ProgressDlg.h
@@ -123,7 +123,7 @@ private:
 
 	bool					m_bAbort;
 	bool					m_bDone;
-	DWORD					m_startTick;
+	ULONGLONG				m_startTick;
 
 	virtual void			DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
 	static UINT				ProgressThreadEntry(LPVOID pVoid);
